+++ docs/history.rst
@@ -4,6 +4,7 @@ History
 Latest
 ------
 - DEP: Drop Python 3.7 support (issue #451)
+- ENH: Add GCPs reading and writing (issue #376)
 
 0.9.1
 ------